requirements/development.txt
# Development requirements
-r base.txt
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0
ipython>=8.0.0
pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.3.0
black>=23.0.0
flake8>=6.0.0
